rename node folder to editor-content
