+++ src/backoffice/editors/media/editor-media.element.ts
@@ -4,7 +4,7 @@ import { customElement, property } from 'lit/decorators.js';
 import type { ManifestEditorView, ManifestWithLoader } from '@umbraco-cms/models';
 import { umbExtensionsRegistry } from '@umbraco-cms/extensions-registry';
 
-import '../shared/node/editor-node.element';
+import '../shared/editor-content/editor-node.element';
 
 @customElement('umb-editor-media')
 export class UmbEditorMediaElement extends LitElement {
@@ -34,7 +34,7 @@ export class UmbEditorMediaElement extends LitElement {
 				type: 'editorView',
 				alias: 'Umb.EditorView.Media.Edit',
 				name: 'Media Editor Edit View',
-				loader: () => import('../shared/node/views/edit/editor-view-node-edit.element'),
+				loader: () => import('../shared/editor-content/views/edit/editor-view-node-edit.element'),
 				weight: 200,
 				meta: {
 					editors: ['Umb.Editor.Media'],
@@ -47,7 +47,7 @@ export class UmbEditorMediaElement extends LitElement {
 				type: 'editorView',
 				alias: 'Umb.EditorView.Media.Info',
 				name: 'Media Editor Info View',
-				loader: () => import('../shared/node/views/info/editor-view-node-info.element'),
+				loader: () => import('../shared/editor-content/views/info/editor-view-node-info.element'),
 				weight: 100,
 				meta: {
 					editors: ['Umb.Editor.Media'],
