+++ src/backoffice/editors/shared/editor-content/editor-node.element.ts
@@ -11,7 +11,7 @@ import { UmbNodeContext } from './node.context';
 import { UmbObserverMixin } from '@umbraco-cms/observable-api';
 import { UmbContextConsumerMixin, UmbContextProviderMixin } from '@umbraco-cms/context-api';
 
-import '../../shared/editor-entity-layout/editor-entity-layout.element';
+import '../editor-entity-layout/editor-entity-layout.element';
 
 // Lazy load
 // TODO: Make this dynamic, use load-extensions method to loop over extensions for this node.
