+++ src/backoffice/editors/shared/editor-content/editor-node.stories.ts
