+++ src/backoffice/editors/shared/editor-content/views/edit/editor-view-node-edit.element.ts
