+++ src/backoffice/editors/shared/editor-content/views/info/editor-view-node-info.element.ts
